Move renderForm to ContactInformationView, working commit
