+++ src/applications/personalization/profile/components/personal-information/addresses/AddressEditView.jsx
@@ -4,25 +4,13 @@ import pickBy from 'lodash/pickBy';
 import ADDRESS_DATA from '~/platform/forms/address/data';
 import { focusElement } from '~/platform/utilities/ui';
 
-import { ADDRESS_POU, FIELD_NAMES, USA } from '@@vap-svc/constants';
+import { ADDRESS_POU, USA } from '@@vap-svc/constants';
 import ContactInformationEditView from '../ContactInformationEditView';
-import CopyMailingAddress from '@@vap-svc/containers/CopyMailingAddress';
-import ContactInfoForm from '@@vap-svc/components/ContactInfoForm';
 
 class AddressEditView extends React.Component {
   componentWillUnmount() {
     focusElement(`#${this.props.fieldName}-edit-link`);
   }
-
-  onInput = (value, schema, uiSchema) => {
-    const newFieldValue = {
-      ...value,
-    };
-    if (newFieldValue['view:livesOnMilitaryBase']) {
-      newFieldValue.countryCodeIso3 = USA.COUNTRY_ISO3_CODE;
-    }
-    this.props.onChangeFormDataAndSchemas(newFieldValue, schema, uiSchema);
-  };
 
   getInitialFormValues = () =>
     this.props.modalData ||
@@ -77,23 +65,6 @@ class AddressEditView extends React.Component {
     );
   };
 
-  renderForm = (formButtons, onSubmit) => (
-    <div>
-      {this.props.fieldName === FIELD_NAMES.RESIDENTIAL_ADDRESS && (
-        <CopyMailingAddress copyMailingAddress={this.copyMailingAddress} />
-      )}
-      <ContactInfoForm
-        formData={this.props.field.value}
-        formSchema={this.props.field.formSchema}
-        uiSchema={this.props.field.uiSchema}
-        onUpdateFormData={this.onInput}
-        onSubmit={onSubmit}
-      >
-        {formButtons}
-      </ContactInfoForm>
-    </div>
-  );
-
   render() {
     return (
       <ContactInformationEditView
@@ -111,11 +82,11 @@ class AddressEditView extends React.Component {
         onDelete={this.props.onDelete}
         onSubmit={this.props.onSubmit}
         refreshTransaction={this.props.refreshTransaction}
-        render={this.renderForm}
         title={this.props.title}
         transaction={this.props.transaction}
         transactionRequest={this.props.transactionRequest}
         uiSchema={this.props.uiSchema}
+        type="address"
       />
     );
   }
