+++ src/applications/personalization/profile/components/personal-information/email-addresses/EmailEditView.jsx
@@ -2,25 +2,9 @@ import React from 'react';
 
 import ContactInformationEditView from '../ContactInformationEditView';
 
-import ContactInfoForm from '@@vap-svc/components/ContactInfoForm';
-
 class EmailEditView extends React.Component {
   getInitialFormValues = () =>
     this.props.data ? { ...this.props.data } : { emailAddress: '' };
-
-  renderForm = (formButtons, onSubmit) => (
-    <>
-      <ContactInfoForm
-        formData={this.props.field.value}
-        formSchema={this.props.field.formSchema}
-        uiSchema={this.props.field.uiSchema}
-        onUpdateFormData={this.props.onChangeFormDataAndSchemas}
-        onSubmit={onSubmit}
-      >
-        {formButtons}
-      </ContactInfoForm>
-    </>
-  );
 
   render() {
     return (
@@ -39,11 +23,11 @@ class EmailEditView extends React.Component {
         onDelete={this.props.onDelete}
         onSubmit={this.props.onSubmit}
         refreshTransaction={this.props.refreshTransaction}
-        render={this.renderForm}
         title={this.props.title}
         transaction={this.props.transaction}
         transactionRequest={this.props.transactionRequest}
         uiSchema={this.props.uiSchema}
+        type="email"
       />
     );
   }
